Reconsitution v2
1. change some abbreviate name to whole name in file "link.py";
2. add an simple readme file;
3. split pcc_emulator and engine into different file;
4. create an file "main.py" for clearly start whole system.
