+++ README.md
@@ -1,0 +1,15 @@
+# Quickly Start
+
+For players, you need to finish the code in file "congestion_control_algorithm.py";
+
+Then, just run the order "python3 main.py".
+
+You will get some output in the path "/output/" and should fix your code according to the output.
+
+# For Detail
+
+## config
+
+## objects
+
+## output
